backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
redis>=5.0.0
fakeredis>=2.20.0
python-jose[cryptography]>=3.3.0
pydantic
pydantic-settings>=2.2.0
python-multipart
passlib[bcrypt]
python-dotenv